x Compilation issue
git-svn-id: svn://svn.code.sf.net/p/mediainfo/code/MediaInfoLib/trunk@6777 45c5ae6f-87cc-4fd0-92ee-e4f023fd80da

+++ Source/MediaInfo/Export/Export_EbuCore.cpp
@@ -736,7 +736,7 @@ Ztring EbuCore_Transform_Video(Ztring &ToReturn, MediaInfo_Internal &MI, size_t 
             AspectRatioString_Pos=AspectRatioString.find(__T('.'));
             if (AspectRatioString_Pos!=(size_t)-1)
             {
-                int64s Denominator=float64_int64s(pow(10, AspectRatioString.size()-AspectRatioString_Pos+1)); //Computing the right denomintor compared to the count of decimals in the value e.g. 1.778 will have a denominator of 1000 (3 digits after the comma)
+                int64s Denominator=float64_int64s(pow((double)10, (int)(AspectRatioString.size()-AspectRatioString_Pos+1))); //Computing the right denomintor compared to the count of decimals in the value e.g. 1.778 will have a denominator of 1000 (3 digits after the comma)
                 factorNumerator=Ztring::ToZtring(AspectRatioString.To_float32()*Denominator, 0);
                 factorDenominator=Ztring::ToZtring(Denominator);
             }
